move annotation functions for easier import w/ epam
