+++ epam/evaluation.py
@@ -5,7 +5,7 @@ import pandas as pd
 import numpy as np
 import bisect
 from epam.utils import pcp_path_of_aaprob_path, load_and_filter_pcp_df
-from scripts.annotate_pcps import get_cdr_fwk_seqs
+from epam.annotate_pcps import get_cdr_fwk_seqs
 from netam.common import SMALL_PROB
 from netam.sequences import (
     AA_STR_SORTED,
